requirements.txt
google-genai
python-dotenv
pyyaml
streamlit
colorama
pydantic
